converted float to double where needed. (#82)
* converted float to double where needed.

* Finished fixing critical bug that caused issues on larger datasets where float was used instead of double.

* Fix: return from min max functions should be double type.
Renamed some variables to match coding standard.

---------

+++ src/view/src/rocprofvis_grid.cpp
@@ -24,17 +24,17 @@ Grid::GetCursorPosition()
 }
 
 void
-Grid::RenderGrid(float min_x, float max_x, float movement, float zoom,
+Grid::RenderGrid(double min_x, double max_x, double movement, float zoom,
                  ImDrawList* draw_list, float scale_x, float v_max_x, float v_min_x,
                  int grid_size)
 {
     ImVec2 cursor_position = ImGui::GetCursorScreenPos();
     ImVec2 content_size    = ImGui::GetContentRegionAvail();
-    float  range           = (v_max_x + movement) - (v_min_x + movement);
+    double range           = (v_max_x + movement) - (v_min_x + movement);
     ImVec2 displaySize     = ImGui::GetIO().DisplaySize;
 
-    float steps = (max_x - min_x) /
-                  (zoom * 20);  // amount the loop which generates the grid iterates by.
+    double steps = (max_x - min_x) /
+                   (zoom * 20);  // amount the loop which generates the grid iterates by.
 
     ImGuiWindowFlags window_flags =
         ImGuiWindowFlags_NoScrollWithMouse | ImGuiWindowFlags_NoMove;
@@ -55,30 +55,30 @@ Grid::RenderGrid(float min_x, float max_x, float movement, float zoom,
         ImVec2 clip_max = ImVec2(child_win.x + child_size.x, child_win.y + child_size.y);
         draw_list->PushClipRect(clip_min, clip_max, true);
 
-        float normalized_start_box = (min_x - (min_x + movement)) * scale_x;
+        double normalized_start_box = (min_x - (min_x + movement)) * scale_x;
         draw_list->AddRectFilled(ImVec2(normalized_start_box, cursor_position.y),
                                  ImVec2(normalized_start_box - 1500.0f,
                                         cursor_position.y + content_size.y - grid_size),
                                  IM_COL32(100, 100, 100, 100));
 
-        float normalized_start_box_end = (max_x - (min_x + movement)) * scale_x;
+        double normalized_start_box_end = (max_x - (min_x + movement)) * scale_x;
         draw_list->AddRectFilled(ImVec2(normalized_start_box_end, cursor_position.y),
                                  ImVec2(normalized_start_box_end + 1500.0f,
                                         cursor_position.y + content_size.y - grid_size),
                                  IM_COL32(100, 100, 100, 100));
 
-        for(float raw_position_points_x = min_x - (steps * 5);
+        for(double raw_position_points_x = min_x - (steps * 5);
             raw_position_points_x < max_x + (steps * 5); raw_position_points_x += steps)
         {
             // loop through min-max and create appropriate number of scale markers with
             // marker value printed at bottom.
 
-            float normalized_start =
+            double normalized_start =
                 (raw_position_points_x - (min_x + movement)) *
                 scale_x;  // this value takes the raw value of the output and converts
                           // them into positions on the chart which is scaled by scale_x
 
-            float normalized_end =
+            double normalized_end =
                 ((raw_position_points_x + steps) - (min_x + movement)) *
                 scale_x;  // this value takes the raw value of the output and converts
                           // them into positions on the chart which is scaled by scale_x
